refactor(data/query): Rename query to series.data
