+++ src/data/series.data.ts
